Provide viewModel to ForYouScreen with Hilt Navigation Compose (read commit notes)
From Android Developers Docs =
"The most important thing to keep in mind when using ViewModel with Compose is that you cannot scope a ViewModel to a composable. This is because a composable is not a ViewModelStoreOwner. Two instances of the same composable in the Composition, or two different composables accessing the same ViewModel type under the same ViewModelStoreOwner would receive the same instance of the ViewModel, which often is not the expected behavior.

To get the benefits of ViewModel in Compose, host each screen in a Fragment or Activity, or use Compose Navigation and use ViewModels in composable functions as close as possible to the Navigation destination. That is because you can scope a ViewModel to Navigation destinations, Navigation graphs, Activities, and Fragments."

So, to provide the viewmodel so it can be used inside the Composable, we need to place the viewModel as highly as possible close to the Navigation
Hierarchy, which Hilt Navigation Compose come in handy. Also we pass the ViewModel into the Composable as a constructor here.

Ref :
https://developer.android.com/topic/libraries/architecture/viewmodel#jetpack-compose
https://developer.android.com/develop/ui/compose/libraries#hilt