MVVM Pattern impl for For You Screen, with State Hoisting and API Calls (read commit notes)
Using real API calls for getting data from API and display it into the list on screen.

The Architecture Pattern used is based on recommended Architecture Journey based on Android Developer Documentation here
https://developer.android.com/topic/architecture
Where the pattern that used here is =
UI Layer (UI -> UI State -> ViewModel) to Data Layer (Repository -> Data Source)
Since this app not include any complex business logic (for now), Domain layer is optional to reduce boilerplate class.

Also, State Hoisting for UI State is implemented based on Android Developer Documentation here
https://developer.android.com/develop/ui/compose/state-hoisting#hoisting-composables
State Hoisting implemented to make UI reusable and easier to manage the state. For this app, the UI state (which is
reflecting API call result) is hoisted into the ViewModel.

And, for collecting UI state in the Composable, the state holder that used in this app utilizing Kotlin Flow which is StateFlow.
Based on Android Developer Documentation here =
https://developer.android.com/kotlin/flow/stateflow-and-sharedflow
https://developer.android.com/kotlin/flow
StateFlow used to collect the UI state with lifecycle aware and to make the app/screen survive configuration changes.
Also with StateFlow, we can ensure that the Flow (from Data Layer) only collected when the UI is subscribed to the Flow with
.stateIn properties and SharingStarted.WhileSubscribed.