Improve State Hoisting Impl for Tourism Place List in ForYouScreen (read commit notes)
Further hoist the state of the main content of this screen which is Item List of Tourism Place by refactoring the List (LazyColumn) into it's own composable component
and hoist the state as a constructor of the composable component.

The state of the TourismPlaceList component is the list item of the Tourism Place.
So, the caller of this component can control the list that will be used, which make it reusable and easy to test.

With this implementation including previous state hoisting attempt, the List of Tourism Place in ForYouScreen still survive configuration changes since
we're already hoist the list (which we're getting from API) to the Screen's ViewModel.

For further ref for State Hoisting =
https://youtu.be/mymWGMy9pYI?si=WMRts7WMxflydh0B