Structuring each layer dependencies (read commit notes)
This app using best practice for modularization as described by Android Developer Docs here =
https://developer.android.com/topic/architecture

So, since this app is a "basic" app which don't need any complex business logic, we dont use "Domain" layer here
to reduce boilerplate classes/operators.

The pattern used here is called "Modern App Architecture", described by Google in their Android Developer Docs, which :
- A reactive and layered architecture.
- Unidirectional Data Flow (UDF) in all layers of the app.
- A UI layer with state holders to manage the complexity of the UI.
- Coroutines and flows.
- Dependency injection best practices.

For Navigation and Dependency Injection (with Hilt), we'll be place it in the central module, which has dependency
to all layer, the :app module. This :app module serve as a "glue" for our layered architecture to Android context and
serve to manage communications across our layered architecture.