Completed Pass Arguments (ID) to Detail Page -> Load Detail from API (+modularized, read commit notes)
Successfully implementing Detail Page retrieving ID from Main Screen and Load Detail Place from API

Modularize Strategy =
Moved DetailPage Screen into it's own feature module, for future improvement for DetailPage Navigation.
Since i already think to add a Favorite Tab to the Top Level Destination (Bottom Navigation Bar) in the future,
so better prepare the Detail Page to it's own feature module since Detail Page soon not only can be accessed from one module,
but two module, which is =
1. For You Module -> Detail Page
2. Favorite Module (soon, next update) -> Detail Page

So, Detail Page is realistic to be moved into it's own feature module so we can reuse this feature across two or more module.

Retrieving ID from Different Module into the Detail Page Module =
Since the main purpose of modularization so each module can be reused (reusability), so each module needed to be
not tightly coupled to another module, or simply we can say the module should be "isolated".
For passing arguments between module (in this case we're passing ID of Tourism Place), we need a third module to "mediate" communication
between module A and module B as described by best practice/recommendation for Modularization in Android Developer Docs below to avoid
cyclic dependencies/tightly coupled modules =
https://developer.android.com/topic/modularization/patterns#communication

In this app, as descibed from Android Dev Docs above, we're mediating communication between ForYou module and DetailPageModule
through our ":app" module where we're already using it for managing our navigation in our app since ":app" module is a central
module which has access to all module in our project.

And, in the target module, which is DetailPage module, we retrieve the ID of Tourism Place via SavedStateHandle with ViewModel
as described by Android Developer Docs here =
https://developer.android.com/develop/ui/compose/navigation#retrieving-complex-data

Work In Progress for Next Update =
Complete Modularization for the rest of feature in this app (Search and Library) Module
and then done.